Rename GithubIssueClient -> GithubApiClient
Rename the Github HTTP client class to be more generic
in prep to add additional methods to it.
