+++ lobby-server/src/main/java/org/triplea/modules/error/reporting/CreateIssueStrategy.java
@@ -10,21 +10,20 @@ import org.triplea.db.dao.error.reporting.ErrorReportingDao;
 import org.triplea.db.dao.error.reporting.InsertHistoryRecordParams;
 import org.triplea.http.client.error.report.ErrorReportRequest;
 import org.triplea.http.client.error.report.ErrorReportResponse;
-import org.triplea.http.client.github.issues.CreateIssueRequest;
-import org.triplea.http.client.github.issues.CreateIssueResponse;
-import org.triplea.http.client.github.issues.GithubIssueClient;
+import org.triplea.http.client.github.CreateIssueRequest;
+import org.triplea.http.client.github.CreateIssueResponse;
+import org.triplea.http.client.github.GithubApiClient;
 
 /** Performs the steps for uploading an error report from the point of view of the server. */
 @Builder
 public class CreateIssueStrategy implements Function<CreateIssueParams, ErrorReportResponse> {
   @Nonnull private final Function<CreateIssueResponse, ErrorReportResponse> responseAdapter;
-  @Nonnull private final GithubIssueClient githubIssueClient;
+  @Nonnull private final GithubApiClient githubApiClient;
   @Nonnull private final ErrorReportingDao errorReportingDao;
 
-  public static CreateIssueStrategy build(
-      final GithubIssueClient githubIssueClient, final Jdbi jdbi) {
+  public static CreateIssueStrategy build(final GithubApiClient githubApiClient, final Jdbi jdbi) {
     return CreateIssueStrategy.builder()
-        .githubIssueClient(githubIssueClient)
+        .githubApiClient(githubApiClient)
         .responseAdapter(new ErrorReportResponseConverter())
         .errorReportingDao(jdbi.onDemand(ErrorReportingDao.class))
         .build();
@@ -50,7 +49,7 @@ public class CreateIssueStrategy implements Function<CreateIssueParams, ErrorRep
 
   private ErrorReportResponse sendRequest(final ErrorReportRequest errorReportRequest) {
     return responseAdapter.apply(
-        githubIssueClient.newIssue(
+        githubApiClient.newIssue(
             CreateIssueRequest.builder()
                 .title(errorReportRequest.getGameVersion() + ": " + errorReportRequest.getTitle())
                 .body(errorReportRequest.getBody())
