+++ lobby-server/src/main/java/org/triplea/modules/error/reporting/ErrorReportResponseConverter.java
@@ -3,7 +3,7 @@ package org.triplea.modules.error.reporting;
 import com.google.common.base.Strings;
 import java.util.function.Function;
 import org.triplea.http.client.error.report.ErrorReportResponse;
-import org.triplea.http.client.github.issues.CreateIssueResponse;
+import org.triplea.http.client.github.CreateIssueResponse;
 
 /**
  * Converts a response from Github.com by our lobby-server into a response object we can send back
