+++ lobby-server/src/test/java/org/triplea/modules/error/reporting/ErrorReportResponseConverterTest.java
@@ -5,7 +5,7 @@ import static org.hamcrest.core.Is.is;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
-import org.triplea.http.client.github.issues.CreateIssueResponse;
+import org.triplea.http.client.github.CreateIssueResponse;
 
 class ErrorReportResponseConverterTest {
